added vector type check

+++ cpp/daal/src/sycl/reducer.cpp
@@ -209,7 +209,7 @@ Reducer::Result Reducer::reduce(const BinaryOp op, Layout vectorsLayout, const U
     Result result(context, resReduce, nVectors, vectors.type(), status);
     DAAL_CHECK_STATUS_RETURN_IF_FAIL(status, Reducer::Result());
 
-    DAAL_ASSERT(vectors.type() == TypeIds::id<float>() || vectors.type() == TypeIds::id<float>());
+    DAAL_ASSERT(vectors.type() == TypeIds::id<float>() || vectors.type() == TypeIds::id<double>());
 
     auto & kernelFactory = context.getClKernelFactory();
